ref: moved schemaDefaults function into schemas/utils
